Adding recipes folder with two recipes
The recipes folder will be used for storing the current recipes and all future incoming.
Currently it contains two recipes, orange chicken (finished) and masala chai (work in progress).
Each recipe has its own webpage with an image, a description, an ingredients list and a step by step guide for preparing the dish.