Fixing error of images not showing
Changed the path from ./img/recipe to ../img/recipe.
The error of the images not showing was caused by trying to find a "img"-folder in the "recipe"-folder.
It should be fixed by correcting the sources path attribute like shown in line 3, "../" will go to the parent folder, in this case "odin-recipes" in which the "img"-folder resides.