Stylizing the homepage of the website
By creating classes ("image", "header", "dish" and "drink") and
assigning them as attributes to the html-tags helped categorizing each
element on the homepage and giving them a uniformed style. (i.e. both
headers "Dishes:" and "Drinks:" share the same font size and
underlining).
Furthermore I split the recipes into the categories "Dishes" and
"Drinks" and created an unordered list for each category. The list
elements contain a referral link to the recipe webpage and an image of
the finished end product of the recipe.
The stylization used (color, font-size, font-weight, etc.) can be found
in the styles.css file.