Stylizing the "Orange Chicken" webpage
I added classes to the html-tags of the "Orange Chicken" webpage in
order to take on the already used style on the homepage.
The order of the classes in style.css had to be adjusted, so the header
class would not be overwritten by the font size of the dish class rather
the other case around.
Furthermore I added another class called "page" with the purpose of
rescaling the image on the webpages with .page.image and indicating that
the image shown is on a recipe webpage and not a small version on the
homepage.