Change layout of homepage and added boxes
I changed the dishes and drinks layout from an unordered list to a grid
like layout. To make it more appealing i packed image and link to the
individual recipe into a button.
The headers are now highlighted in boxes and do not stick to the wall,
they have a margin and a padding to avoid sticking to borders or walls
(wall = border of the browser).
I also replaced the nested <p> <img> combination with "display: block".
It will make it easier creating buttons for future recipes.